recommender_service/requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
requests
